+++ x/profiles/keeper/alias_functions.go
@@ -33,6 +33,30 @@ func (k Keeper) GetProfiles(ctx sdk.Context) []*types.Profile {
 	return profiles
 }
 
+// --------------------------------------------------------------------------------------------------------------------
+
+// IterateUserIncomingDTagTransferRequests iterates over all the DTag transfer request made to the given user
+// and performs the provided function
+func (k Keeper) IterateUserIncomingDTagTransferRequests(
+	ctx sdk.Context, user string, fn func(index int64, dTagTransferRequest types.DTagTransferRequest) (stop bool),
+) {
+	store := ctx.KVStore(k.storeKey)
+	iterator := sdk.KVStorePrefixIterator(store, types.IncomingDTagTransferRequestsPrefix(user))
+	defer iterator.Close()
+
+	i := int64(0)
+	for ; iterator.Valid(); iterator.Next() {
+		request := types.MustUnmarshalDTagTransferRequest(k.cdc, iterator.Value())
+		stop := fn(i, request)
+		if stop {
+			break
+		}
+		i++
+	}
+}
+
+// --------------------------------------------------------------------------------------------------------------------
+
 // IterateRelationships iterates through the relationships and perform the provided function
 func (k Keeper) IterateRelationships(ctx sdk.Context, fn func(index int64, relationship types.Relationship) (stop bool)) {
 	store := ctx.KVStore(k.storeKey)