+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Added the possibility to verify a profile with an external application ([#472](https://github.com/desmos-labs/desmos/issues/472)) 
 - Improved the posts query ([#499](https://github.com/desmos-labs/desmos/issues/499))
 - Added the post comments query ([#510](https://github.com/desmos-labs/desmos/pull/510))
+- Added the ability to paginate incoming DTag transfer requests ([#519](https://github.com/desmos-labs/desmos/pull/519))
 
 # Version 0.16.3
 ## Changes